lego-vision/requirements.txt
# Azure Custom Vision SDK for object detection
azure-cognitiveservices-vision-customvision>=3.1.0
azure-identity>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.2.0
